feat(rome_js_formatter): Binary like object literal member expression
