+++ crates/rome_js_formatter/src/utils/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod array;
 mod binary_like_expression;
 mod format_conditional;
+mod object;
 mod simple;
 pub mod string_utils;
 
@@ -12,6 +13,10 @@ use crate::prelude::*;
 pub(crate) use binary_like_expression::{format_binary_like_expression, JsAnyBinaryLikeExpression};
 pub(crate) use format_conditional::{format_conditional, Conditional};
 pub(crate) use member_chain::format_call_expression;
+pub(crate) use object::{
+    is_break_after_colon, property_object_member_layout, write_member_name,
+    PropertyObjectMemberLayout,
+};
 use rome_formatter::{normalize_newlines, write, Buffer, VecBuffer};
 use rome_js_syntax::suppression::{has_suppressions_category, SuppressionCategory};
 use rome_js_syntax::JsSyntaxKind::JS_STRING_LITERAL;
